fixed bug in XSLT
Vocalized Syriac names were sometimes being tokenized on "\.\s" and
sometimes on ".\s" (erroneously).  I fixed the XSLT: now all the
tokenize() functions on vocalized Syriac names are tokenizing on "\.\s"
as they are supposed to.  This explains why vocalized Syriac names with
multiple words in the name were being given an @xml:lang="en" instead of
"syr-Syrj" and lacked a @source attribute.